backend/ai_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers>=4.35.0
torch>=2.0.0
pydantic==2.5.0
python-multipart==0.0.6
google-generativeai>=0.3.0
pillow>=10.0.0